requirements.txt
aiogram==3.15.0
qrcode[pil]==8.0
Pillow==10.4.0
aiosqlite==0.20.0
python-dotenv==1.0.1
numpy==1.26.4
